Interface and partial implementation of the Moving_game_object and Moving_game_object_manager classes.
The Moving_game_object class will be used to represent object that will appear to be moving in the game.  This includes the skier and gates, but not things such as the score, or other HUD-type information.  The Moving_game_object_manager will be used to manage moving game objects such that one is focused while the others move appropriately proportionally to that focused object.