Really brutal first attempt at having all sprites use the same sheet, but hey, it worked.
A lot still needs to be added in this area.  Primarily, the location of the global object is within the game class, although it might be better to have in the Sprites namespace.  Error handling is also pretty much just not implemented. Most of this should be fixed in my next commit.